small changes (thirteenth commit)

+++ FFmpeg.cs
@@ -12,6 +12,15 @@ namespace TimelapseApp
 
         public static Task Record(string sourceLink, string outputPath, int recordingTime)
         {
+            if (!sourceLink.IsRtspLinkValid())
+                return Task.FromResult(new InvalidDataException("RTSP-link is invalid"));
+
+            if (recordingTime < 1)
+            {
+                $"[FFmpeg.Record()]: The video time index {recordingTime} is lower than it could be".Message();
+                return Task.FromException(new IndexOutOfRangeException($"The acceleration index {recordingTime} is lower than it could be"));
+            }
+
             bool localTimeChecked = Config.GetTimestampChecked();
 
             string localTimeString = "-vf drawtext=\"font=Arial:fontsize=70:fontcolor=white:x=0:y=0:text='%{localtime\\:%b %d %Y %H\\\\\\:%M\\\\\\:%S}':box=1:boxcolor=black@0.4:boxborderw=5\" ";
@@ -20,24 +29,63 @@ namespace TimelapseApp
                 $"{(localTimeChecked ? localTimeString : "")}" +
                 $"-t {recordingTime} {outputPath}";
 
-            using Process record = Process.Start(new ProcessStartInfo("ffmpeg", args));
+            using Process record = new();
+            record.StartInfo = new("ffmpeg", args);
+            record.Start();
             record.WaitForExit();
+
+            if (!File.Exists(outputPath))
+                $"[FFmpeg.Record()]: Video was not recorded".Message();
 
             return Task.FromResult(record);
         }
 
-        public static Task Accelerate(string sourcePath, string resultPath, int acceleration)
+        public static Task Accelerate(string sourcePath, string outputPath, int acceleration)
         {
-            string args = $"-y -hide_banner -i {sourcePath} -c:v libx264 -r 25 -vf setpts=PTS/{acceleration} -an {resultPath}";
+            if (!File.Exists(sourcePath))
+            {
+                $"[FFmpeg.Accelerate()]: File {sourcePath} was not found".Message();
+                return Task.FromException(new FileNotFoundException("File was not found", sourcePath));
+            }
 
-            using Process accelerate = Process.Start(new ProcessStartInfo("ffmpeg", args));
+            if (acceleration < 1)
+            {
+                $"[FFmpeg.Accelerate()]: The acceleration index {acceleration} is lower than it could be".Message();
+                return Task.FromException(new IndexOutOfRangeException($"The acceleration index {acceleration} is lower than it could be"));
+            }
+
+            string args = $"-y -hide_banner -i {sourcePath} -c:v libx264 -r 24 -vf setpts=PTS/{acceleration} -an {outputPath}";
+
+            using Process accelerate = new();
+            accelerate.StartInfo = new("ffmpeg", args);
+            accelerate.Start();
             accelerate.WaitForExit();
+
+            if (!File.Exists(outputPath))
+                $"[FFmpeg.Accelerate()]: Video was not accelerated".Message();
 
             return Task.FromResult(accelerate);
         }
 
-        public static Task Concat(List<string> sourseVideos, string resultPath)
+        public static Task Concat(List<string> sourseVideos, string outputPath)
         {
+            if (sourseVideos.Count == 0)
+            {
+                $"[FFmpeg.Concat()]: There is no videos to concat".Message();
+                return Task.FromException(new FileNotFoundException("File was not found", outputPath));
+            }
+            else
+            {
+                foreach (string sourseVideo in sourseVideos)
+                {
+                    if (!File.Exists(sourseVideo))
+                    {
+                        $"[FFmpeg.Concat()]: File {sourseVideo} was not found".Message();
+                        return Task.FromException(new FileNotFoundException("File was not found", sourseVideo));
+                    }
+                }
+            }
+
             string concatFilePath = Path.Combine(Temp.Path, "concat.txt");
             using (StreamWriter sw = new(concatFilePath))
             {
@@ -45,34 +93,59 @@ namespace TimelapseApp
                     sw.WriteLine($"file '{video}'");
             }
 
-            string args = $"-y -hide_banner -f concat -safe 0 -i {concatFilePath} -c copy {resultPath}";
+            string args = $"-y -hide_banner -f concat -safe 0 -i {concatFilePath} -c copy {outputPath}";
 
-            using Process concat = Process.Start(new ProcessStartInfo("ffmpeg", args));
+            using Process concat = new();
+            concat.StartInfo = new("ffmpeg", args);
+            concat.Start();
             concat.WaitForExit();
 
-            File.Delete(concatFilePath);
+            if (File.Exists(outputPath))
+            {
+                try
+                {
+                    File.Delete(concatFilePath);
+                }
+                catch (Exception ex)
+                {
+                    ("[FFmpeg.Concat.File.Delete()]: " + ex.Message).Message();
+                }
+            }
+            else "[FFmpeg.Concat()]: The videos were not concated".Message();
 
             return Task.FromResult(concat);
         }
 
         public static Task Repair(string sourcePath)
         {
-            string repairedVideoPath = Path.Combine(Temp.Path, "repairedVideo.mp4");
-            string args = $"-y -hide_banner -i {sourcePath} -ss 0 -c:v copy -an -copyts -start_number 0 {repairedVideoPath}";
+            if (!File.Exists(sourcePath))
+            {
+                $"[FFmpeg.Repair()]: File {sourcePath} was not found".Message();
+                return Task.FromException(new FileNotFoundException("File was not found", sourcePath));
+            }
 
-            using Process repair = Process.Start(new ProcessStartInfo("ffmpeg", args));
+            string repairedVideoPath = Path.Combine(Temp.Path, "repairedVideo.mp4");
+            string args = $"-y -hide_banner -i {sourcePath} -c:v copy -an -copyts -start_number 0 {repairedVideoPath}";
+
+            using Process repair = new();
+            repair.StartInfo = new("ffmpeg", args);
+            repair.Start();
             repair.WaitForExit();
 
-            try
+            if (File.Exists(repairedVideoPath))
             {
-                File.Move(repairedVideoPath, sourcePath, true);
+                try
+                {
+                    File.Move(repairedVideoPath, sourcePath, true);
+                }
+                catch (Exception ex)
+                {
+                    ("[FFmpeg.Repair.File.Move()]: " + ex.Message).Message();
+                }
             }
-            catch (Exception ex)
-            {
-                ("[FFmpeg.Repair.File.Move()]: " + ex.Message).Message();
-            }
+            else "[FFmpeg.Repair()]: The video was not repaired".Message();
 
-            return Task.FromResult(repair);
+            return Task.FromResult(repair.ExitCode);
         }
     }
 }
